+++ packages/builder/builder-shared/src/types/plugin.ts
@@ -1,0 +1,11 @@
+export type PluginStore = {
+  readonly plugins: BuilderPlugin[];
+  addPlugins: (plugins: BuilderPlugin[]) => void;
+  removePlugins: (pluginNames: string[]) => void;
+  isPluginExists: (pluginName: string) => boolean;
+};
+
+export type BuilderPlugin<API = any> = {
+  name: string;
+  setup: (api: API) => Promise<void> | void;
+};