+++ packages/builder/builder-webpack-provider/src/index.ts
@@ -1,0 +1,28 @@
+export { builderWebpackProvider } from './provider';
+export type {
+  BuilderPluginAPI,
+
+  // Config Types
+  DevConfig,
+  HtmlConfig,
+  OutputConfig,
+  SourceConfig,
+  BuilderConfig,
+  SecurityConfig,
+  PerformanceConfig,
+  ExperimentsConfig,
+
+  // Third Party Types
+  webpack,
+  WebpackChain,
+  WebpackConfig,
+  CSSLoaderOptions,
+  CssExtractOptions,
+  LessLoaderOptions,
+  SassLoaderOptions,
+  HTMLPluginOptions,
+  StyleLoaderOptions,
+  AutoprefixerOptions,
+  TerserPluginOptions,
+  PostCSSLoaderOptions,
+} from './types';