+++ packages/builder/builder-webpack-provider/src/stub/context.ts
@@ -1,0 +1,6 @@
+import { createPrimaryContext } from '../core/createContext';
+import type { StubBuilderOptions } from './builder';
+
+export function createStubContext(options: Required<StubBuilderOptions>) {
+  return createPrimaryContext(options, options.builderConfig);
+}