+++ packages/toolkit/types/server/devServer.d.ts
@@ -15,9 +15,6 @@ export type DevServerOptions = {
     path?: string;
     port?: string;
     host?: string;
-    logging?: string;
-    overlay?: boolean;
-    progress?: boolean;
   };
   devMiddleware?: {
     writeToDisk: boolean | ((filename: string) => boolean);
@@ -34,5 +31,19 @@ export type DevServerOptions = {
   liveReload?: boolean;
   /** Whether to enable https. */
   https?: DevServerHttpsOptions;
+  /** see https://github.com/bripkens/connect-history-api-fallback */
+  historyApiFallback?:
+    | boolean
+    | {
+        index?: string;
+        verbose?: boolean;
+        logger?: typeof console.log;
+        htmlAcceptHeaders?: string[];
+        disableDotRule?: true;
+        rewrites?: Array<{
+          from: RegExp;
+          to: string | RegExp | function;
+        }>;
+      };
   [propName: string]: any;
 };