+++ scripts/prebundle/src/constant.ts
@@ -188,13 +188,17 @@ export const TASKS: TaskConfig[] = [
     ],
   },
   {
-    packageDir: 'builder/webpack-builder',
-    packageName: '@modern-js/webpack-builder',
+    packageDir: 'builder/builder-webpack-provider',
+    packageName: '@modern-js/builder-webpack-provider',
     dependencies: [
       'serialize-javascript',
       'open',
       'tapable',
       'webpack-merge',
+      {
+        name: 'postcss-pxtorem',
+        ignoreDts: true,
+      },
       {
         name: 'webpack-5-chain',
         externals: {
@@ -372,7 +376,7 @@ export const TASKS: TaskConfig[] = [
         name: 'url-loader',
         ignoreDts: true,
         externals: {
-          'schema-utils': 'schema-utils',
+          'schema-utils': '../schema-utils3',
           'loader-utils': '../loader-utils2',
           'mime-types': '@modern-js/utils/mime-types',
         },
@@ -791,13 +795,13 @@ export const TASKS: TaskConfig[] = [
   },
   {
     packageDir: 'builder/plugin-esbuild',
-    packageName: '@modern-js/webpack-builder-plugin-esbuild',
+    packageName: '@modern-js/builder-plugin-esbuild',
     dependencies: [
       {
         name: 'esbuild-loader',
         ignoreDts: true,
         externals: {
-          '/^webpack(/.*)/': '@modern-js/webpack-builder/webpack$1',
+          '/^webpack(/.*)/': '@modern-js/builder-webpack-provider/webpack$1',
         },
         afterBundle(task) {
           const dtsFiles = glob.sync(join(task.depPath, 'dist', '*.d.ts'), {