+++ tests/integration/webpack-builder/scripts/watch.ts
@@ -1,0 +1,9 @@
+import { createBuilder } from './shared';
+
+(async function main() {
+  const builder = await createBuilder();
+  await builder.build({
+    watch: true,
+    mode: 'development',
+  });
+})();