+++ packages/cli/babel-preset-base/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/babel-preset-base
 
+## 1.19.0
+
+### Patch Changes
+
+- @modern-js/utils@1.19.0
+
 ## 1.18.1
 
 ### Patch Changes
@@ -62,7 +68,7 @@
 
 ### Minor Changes
 
-- 7b9067f: add babel plugin for webpack-builder
+- 7b9067f: add babel plugin for builder
 
 ### Patch Changes