+++ packages/runtime/plugin-runtime/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @modern-js/runtime
 
+## 1.19.0
+
+### Patch Changes
+
+- @modern-js/webpack@1.19.0
+- @modern-js/plugin@1.19.0
+- @modern-js/utils@1.19.0
+
 ## 1.18.1
 
 ### Patch Changes